Update Address.php
Addresses returned by the getOrder endpoint are documented as requiring an OrderAddress Address name parameter, but responses are not conforming to that spec.

Currently, requests to JsonSerialize on an Orders object will fail when the getName method is called due to the return type. Adjusting the return to be an optional string.